adding Screen from different news
