+++ lib/Constants/constants.dart
@@ -1,0 +1,12 @@
+String topNewsUrl =
+    "https://newsapi.org/v2/top-headlines?country=us&apiKey=eaa9c42540ee41cd9350ea0074df4356";
+String bussinessUrl =
+    "https://newsapi.org/v2/top-headlines?country=us&category=business&apiKey=eaa9c42540ee41cd9350ea0074df4356";
+String entertainmentUrl =
+    "https://newsapi.org/v2/top-headlines?country=us&category=entertainment&apiKey=eaa9c42540ee41cd9350ea0074df4356";
+String sportsUrl =
+    "https://newsapi.org/v2/top-headlines?country=us&category=sports&apiKey=eaa9c42540ee41cd9350ea0074df4356";
+String scienceUrl =
+    "https://newsapi.org/v2/top-headlines?country=us&category=science&apiKey=eaa9c42540ee41cd9350ea0074df4356";
+String technologyUrl =
+    "https://newsapi.org/v2/top-headlines?country=us&category=technology&apiKey=eaa9c42540ee41cd9350ea0074df4356";
