+++ lib/Screens/HomePage.dart
@@ -1,247 +1,88 @@
-import 'dart:convert';
-
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/Screens/newsScreen.dart';
-import 'package:flutter_application_1/widgets/navBar.dart';
-import 'package:http/http.dart' as http;
+import 'package:flutter_application_1/Screens/ScienceNews.dart';
+import 'package:flutter_application_1/Screens/Technology_screen.dart';
+import 'package:flutter_application_1/Screens/TopNews.dart';
+import 'package:flutter_application_1/Screens/entertainment.dart';
+import 'package:flutter_application_1/Screens/sports.dart';
+
+import 'Business_screen.dart';
 
 class HomeScreen extends StatefulWidget {
-  const HomeScreen({super.key});
+  const HomeScreen({Key? key}) : super(key: key);
 
   @override
   State<HomeScreen> createState() => _HomeScreenState();
 }
 
 class _HomeScreenState extends State<HomeScreen> {
-  double height = 50;
-  double width = 50;
-  double height1 = 50;
-  double width1 = 300;
-  Color border_color = Colors.white;
-  Color color = Colors.transparent;
+  final List<String> navBarItem = [
+    "Business News",
+    "Technology News",
+    "Science News",
+    "Sports News",
+    "Top News",
+    "Entertainment News"
+  ];
 
-  String? stringResponse;
-  Map? mapResponse;
-  List listResponse = [];
-
-  Future apiCall() async {
-    http.Response response;
-    String url =
-        "https://newsapi.org/v2/everything?q=tesla&from=2023-07-10&sortBy=publishedAt&apiKey=eaa9c42540ee41cd9350ea0074df4356";
-
-    response = await http.get(Uri.parse(url));
-
-    if (response.statusCode == 200) {
-      setState(() {
-        stringResponse = response.body;
-        mapResponse = json.decode(response.body);
-        listResponse = mapResponse!["articles"];
-      });
-    } else {
-      debugPrint("Error in loading!!!");
+  void _navigateToScreen(String item) {
+    if (item == "Business News") {
+      Navigator.of(context).push(
+        MaterialPageRoute(builder: (context) => const BusinessScreen()),
+      );
+    }
+    if (item == "Technology News") {
+      Navigator.of(context).push(
+        MaterialPageRoute(builder: (context) => const TechnologyScreen()),
+      );
+    }
+    if (item == "Science News") {
+      Navigator.of(context).push(
+        MaterialPageRoute(builder: (context) => const ScienceScreen()),
+      );
+    }
+    if (item == "Sports News") {
+      Navigator.of(context).push(
+        MaterialPageRoute(builder: (context) => const SportsScreen()),
+      );
+    }
+    if (item == "Top News") {
+      Navigator.of(context).push(
+        MaterialPageRoute(builder: (context) => const TopScreen()),
+      );
+    }
+    if (item == "Entertainment News") {
+      Navigator.of(context).push(
+        MaterialPageRoute(builder: (context) => const EntertainmentScreen()),
+      );
     }
   }
 
-  @override
-  void initState() {
-    apiCall();
-    super.initState();
-  }
-
+  // if(item == ){}
+  // }
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        iconTheme: const IconThemeData(color: Colors.deepPurple),
-        backgroundColor: Colors.white,
-        elevation: 0,
-        leading: const Icon(Icons.menu),
-        actions: const [Icon(Icons.notifications)],
-        title: const Text(
-          "Taza Khabar",
-          style: TextStyle(color: Colors.deepPurple),
-        ),
+        title: const Text('List of News'),
       ),
-      body: SingleChildScrollView(
-        child: Column(
-          children: [
-            Row(
-              children: [
-                AnimatedContainer(
-                  duration: const Duration(milliseconds: 20),
-                  height: height1,
-                  width: width1,
-                  child: const Padding(
-                    padding: EdgeInsets.all(8.0),
-                    child: Text(
-                      "Breaking News",
-                      style:
-                          TextStyle(fontSize: 30, fontStyle: FontStyle.italic),
-                    ),
-                  ),
-                ),
-                AnimatedContainer(
-                  duration: const Duration(milliseconds: 20),
-                  height: height,
-                  width: width,
-                  decoration: BoxDecoration(
-                      border: Border.all(color: border_color, width: 0.0),
-                      borderRadius: BorderRadius.circular(10),
-                      color: color),
-                  child: Padding(
-                    padding: const EdgeInsets.all(8.0),
-                    child: Row(
-                      children: [
-                        InkWell(
-                          onTap: () {
-                            color = color;
-
-                            Colors.black;
-                            height = 0.0;
-                            width = 0.0;
-                            setState(() {});
-                          },
-                          onDoubleTap: () {
-                            color = color;
-
-                            Colors.white;
-                            height = 0.0;
-                            width = 0.0;
-                            setState(() {});
-                          },
-                          child: const Icon(Icons.search),
-                        ),
-                        const Expanded(
-                          child: TextField(
-                            textInputAction: TextInputAction.search,
-                            decoration:
-                                InputDecoration(border: InputBorder.none),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(
-              height: 10,
-            ),
-            const NavBar(),
-            Container(
-              height: MediaQuery.of(context).size.height * 0.3,
-              width: MediaQuery.of(context).size.width * 0.9,
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(8),
-                color: Colors.amber,
-                image: const DecorationImage(
-                    image: NetworkImage(
-                        "https://cleantechnica.com/files/2023/08/2023_06_Electrify_DC_JoseValdezPhoto-2305-meguiars-aj-velasco-tesla-model-3-best-of-show-scaled.jpg"),
-                    fit: BoxFit.cover),
-              ),
-            ),
-            Container(
-              height: MediaQuery.of(context).size.height,
-              width: MediaQuery.of(context).size.width * 0.9,
-              child: ListView.builder(
-                shrinkWrap: true,
-                physics: NeverScrollableScrollPhysics(),
-                itemCount: listResponse.length,
-                itemBuilder: (context, index) {
-                  return InkWell(
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                          builder: (context) => NewsScreen(
-                              img: listResponse[index]['urlToImage'],
-                              description: listResponse[index]['description'],
-                              author: listResponse[index]['author'],
-                              title: listResponse[index]['title']),
-                        ),
-                      );
-                    },
-                    child: Container(
-                      height: 100,
-                      width: MediaQuery.of(context).size.width * 0.9,
-                      decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(10),
-                          color: Colors.white,
-                          border:
-                              Border.all(color: Colors.deepPurple, width: 0.5)),
-                      child: Row(
-                        children: [
-                          Container(
-                            width: 100,
-                            height: 100,
-                            decoration: BoxDecoration(
-                                color: Colors.blue,
-                                borderRadius: BorderRadius.circular(10.0),
-                                image: DecorationImage(
-                                    image: NetworkImage(
-                                        listResponse[index]['urlToImage']),
-                                    fit: BoxFit.cover)),
-                          ),
-                          Flexible(
-                            child: Column(
-                              children: [
-                                Padding(
-                                  padding: const EdgeInsets.only(
-                                      left: 8.0, right: 8.0, top: 20),
-                                  child: RichText(
-                                    overflow: TextOverflow.ellipsis,
-                                    text: TextSpan(
-                                      text: listResponse[index]['description'],
-                                    ),
-                                  ),
-                                ),
-                                const Padding(
-                                  padding: EdgeInsets.only(left: 10, top: 20),
-                                  child: Row(
-                                    children: [
-                                      Text(
-                                        "Source",
-                                        style: TextStyle(fontSize: 12),
-                                      ),
-                                      Text(
-                                        "BBC",
-                                        style: TextStyle(fontSize: 12),
-                                      ),
-                                      Text(
-                                        "10 min ago",
-                                        style: TextStyle(fontSize: 12),
-                                      ),
-                                      SizedBox(
-                                        height: 10,
-                                      ),
-                                      Icon(
-                                        Icons.favorite,
-                                        size: 12,
-                                        color: Colors.deepPurple,
-                                      ),
-                                      Padding(
-                                        padding: EdgeInsets.only(left: 10.0),
-                                        child: Icon(
-                                          Icons.share,
-                                          size: 12,
-                                          color: Colors.deepPurple,
-                                        ),
-                                      )
-                                    ],
-                                  ),
-                                )
-                              ],
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  );
-                },
-              ),
-            )
-          ],
+      body: GridView.builder(
+        gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+          crossAxisCount: 2,
         ),
+        itemCount: navBarItem.length,
+        itemBuilder: (context, index) {
+          return GestureDetector(
+            onTap: () {
+              _navigateToScreen(navBarItem[index]);
+            },
+            child: Card(
+              elevation: 3,
+              child: Center(
+                child: Text(navBarItem[index]),
+              ),
+            ),
+          );
+        },
       ),
     );
   }
