+++ lib/Screens/Technology_screen.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_application_1/Constants/constants.dart';
+import 'package:flutter_application_1/Screens/Detail.dart.dart';
+import 'package:flutter_application_1/Screens/HomePage.dart';
+
+class TechnologyScreen extends StatefulWidget {
+  const TechnologyScreen({super.key});
+
+  @override
+  State<TechnologyScreen> createState() => _TechnologyScreenState();
+}
+
+class _TechnologyScreenState extends State<TechnologyScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        backgroundColor: Colors.amber,
+        appBar: AppBar(
+          actions: [
+            IconButton(
+                onPressed: () {
+                  Navigator.of(context).pushReplacement(MaterialPageRoute(
+                      builder: (context) => const HomeScreen()));
+                },
+                icon: const Icon(Icons.arrow_back))
+          ],
+        ),
+        body: DetailedScreen(url: technologyUrl, text: "Technology Screen"));
+  }
+}
