+++ lib/Screens/newsScreen.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_application_1/Screens/Detail.dart.dart';
 import 'package:flutter_application_1/Screens/HomePage.dart';
 
 class NewsScreen extends StatefulWidget {
@@ -21,17 +22,19 @@ class _NewsScreenState extends State<NewsScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(actions: [
-        IconButton(
-            onPressed: () {
-              Navigator.of(context).pushReplacement(
-                MaterialPageRoute(
-                  builder: (context) => const HomeScreen(),
-                ),
-              );
-            },
-            icon: const Icon(Icons.arrow_back))
-      ]),
+      appBar: AppBar(
+        actions: [
+          IconButton(
+              onPressed: () {
+                Navigator.of(context).pushReplacement(
+                  MaterialPageRoute(
+                    builder: (context) => const HomeScreen(),
+                  ),
+                );
+              },
+              icon: const Icon(Icons.arrow_back))
+        ],
+      ),
       body: SingleChildScrollView(
         child: Column(children: [
           Container(
